6.0.7 Changelog:
ModMii now uses TMDedit by person66 to patch TMDs for future d2x beta cIOSs. This will allow davebaol to add or remove modules in future cIOSs without requiring further changes made to ModMii in order to support it.

Updated Casper to v0.3, thanks giantpune!

Updated name of WiiFlow to WiiFlow-Mod

ModMii no longer needlessly downloads 00000001.app from IOS60 when using the latest sneek installer v0.7a. This version of the sneek installer is used when building neek2o rev70+ or neek rev186+.

Added check to Casper2Bootmii for the temp\DownloadQueues folder and creates it if it's missing. This will fix errors that some may have encountered if that folder was missing.

Fixed error when saving an advanced cIOS to a download queue when its slot was not changed.

Other minor changes.